requirements.txt
pyautogui
Pillow
img2pdf
pygetwindow
opencv-python
numpy
